-overwrite: seek instead of relying on re-reading the partition table
This makes the packer more flexible with regards to permissions: write
permissions for the target device are now sufficient.

The downside is that the SD card needs to be unplugged and re-plugged before the
persistent file system can be created (if desired).

related to https://github.com/gokrazy/tools/issues/9