Added proof of concept implementation for TLS (gokrazy/gokrazy#7)
This is the packer part of a basic implementation of TLS in gokrazy.
New Paramters:

-tls=self-signed
If set, the packer generates and signs a RSA4096 certificate.
The certificate and the key are stored in
~/.config/gokrazy/<hostname>/{cert,key}.pem
-tls=<path>[,<path>]
If set, the packer uses a certificate (+key) specified by the user.

General behavior:
The specified certificate/key is placed in /etc/ssl/{web,web_key}.pem on
the PI.

The host specific configuration directory is used to verify the host's
identity and to perfom a ssl stripping test when updating gokrazy via
http.
If -tls=<path>[,<path>] is given, the updater adds the specified
certificate to its CertPool.