fix sbom generation inconsistency
prior to this commit SBOMs would have inconsistency in their hashing on
the configuration file.

The representation of the config file in fact would differ at certain
stages of the gok commands lifecycle, where at packer running time,
an extra InternalCompatibilityFlag, Sudo, would be added in memory
(while that not being the case at `gok sbom` time),
resulting in a differing config and as such differing SBOM hashes, same
goes for the differing pointer addresses that were skew the hashing
results.

This is now fixed by removing the InternalCompatibilityFlag set (using
an accessor) from the
config prior to its hashing, as well as converting the config into a
string before performing the hash computation, to avoid differing pointer problems.